+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -132,7 +132,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-30T08:39:00.224Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-30T09:07:30.737Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")