+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMYYBBHYT09Y94VARR0B73GC"
+        "01KMYZZK7S8DK3PBFRMQ3N8ZCB"
     )
     r"""The unique identifier of the data source"""