+++ packages/orq-rc/src/orq_ai_sdk/models/createeval_guardrail_config_evals_response_200_applicationjson_responsebody_4_2_type.py
@@ -1249,9 +1249,9 @@ class Typescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-03-30T08:39:02.541Z"
+    created: Optional[str] = "2026-03-30T09:07:33.148Z"
 
-    updated: Optional[str] = "2026-03-30T08:39:02.541Z"
+    updated: Optional[str] = "2026-03-30T09:07:33.148Z"
 
     guardrail_config: Optional[
         CreateEvalResponseBodyEvalsResponse200ApplicationJSONGuardrailConfig
@@ -1445,9 +1445,9 @@ class Ragas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-03-30T08:39:02.541Z"
+    created: Optional[str] = "2026-03-30T09:07:33.148Z"
 
-    updated: Optional[str] = "2026-03-30T08:39:02.541Z"
+    updated: Optional[str] = "2026-03-30T09:07:33.148Z"
 
     guardrail_config: Optional[
         CreateEvalResponseBodyEvalsResponse200GuardrailConfig
@@ -2140,9 +2140,9 @@ class CreateEvalResponseBodyFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-03-30T08:39:02.541Z"
+    created: Optional[str] = "2026-03-30T09:07:33.148Z"
 
-    updated: Optional[str] = "2026-03-30T08:39:02.541Z"
+    updated: Optional[str] = "2026-03-30T09:07:33.148Z"
 
     guardrail_config: Optional[CreateEvalResponseBodyEvalsResponseGuardrailConfig] = (
         None