+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -1903,7 +1903,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KMYYBB578HRB3PA4XGEZVW79"
+    id: Optional[str] = "01KMYZZJVV3TV716NDAKBQBPV7"
 
     description: Optional[str] = None