+++ packages/orq-rc/RELEASES.md
@@ -6699,3 +6699,13 @@ Based on:
 - [python v4.7.0-rc.29] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.29] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.29 - packages/orq-rc
+
+## 2026-03-30 09:10:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.3 (2.869.25) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.30] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.30] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.30 - packages/orq-rc